test: Update ostree-ng.sh to cover bug BZ#1951087
Update the first blueprint to use "normal" kernel instead of to
use "rt" kernel. Then upgrade to "rt" kernel.

To work with this bug fix, we need osbuild fix and have to pin to
osbuild latest commit.

The customizations.user is removed from upgrade blueprint to work
with this bug fix. But it's for RHEL 8.5 only.