tests: enable vmware test on RHEL-9
These changes enable the use of weldr-client with the test and also
there was a need to change the vm.create command a little bit. Without
the increased CPU and RAM RHEL-9 would not boot and a network adapter
had to be specified to use a newer one because the default one does not
work with RHEL-9.

(cherry picked from commit 5350b42b1df7894e02331b4fb06ed5a364275a1a)