customizations: add support for fs sizes in MB and GB
It is not convenient to specify filesystem sizes in bytes every time.
Add support for parsing string with mMgG specified.

This approach has a downside, that now the filesystem size is of type
String and therefore all of the references must be strings which changes
the original behavior.

Alternative approach is to define a custom type and enable parsing of
both number and string.

TODO: input validation in Weldr API